+++ tico/quantization/wrapq/examples/llama/quantize_llama_mlp.py
@@ -29,7 +29,7 @@ from tico.quantization.wrapq.wrappers.llama.quant_mlp import QuantLlamaMLP
 from tico.utils.utils import SuppressWarning
 
 name = "Maykeye/TinyLLama-v0"
-model = AutoModelForCausalLM.from_pretrained(name)
+model = AutoModelForCausalLM.from_pretrained(name, dtype=torch.float32)
 tokenizer = AutoTokenizer.from_pretrained(name)
 model.eval()
